+++ stor/dx.py
@@ -165,6 +165,12 @@ class DXPath(OBSPath):
         parts = self._get_parts()
         return parts[0] if len(parts) > 0 and parts[0] else None
 
+    def joinpath(self, *others):
+        """Wrapper around base joinpath function which converts the first part to normpath
+        before joining with others.
+        """
+        return self.path_class(self.path_module.join(self.normpath(), *others))
+
     def temp_url(self, lifetime=300, filename=None):
         """Obtains a temporary URL to a DNAnexus data-object.
 
@@ -1176,6 +1182,13 @@ class DXVirtualPath(DXPath):
             return norm_pth.normpath()
         return norm_pth
 
+    def splitpath(self):
+        """Wrapper around base splitpath function which calls splitpath on the normpath of self
+        """
+        path_to_split = self.normpath()
+        parent, child = self.path_module.split(path_to_split)
+        return self.path_class(parent), child
+
 
 class DXCanonicalPath(DXPath):
     """Represents fully canonicalized DNAnexus paths:
@@ -1222,3 +1235,14 @@ class DXCanonicalPath(DXPath):
 
     def normpath(self):
         return self.path_class(self.drive + self.project + ':' + (self.resource or ''))
+
+    def splitpath(self):
+        """Wrapper around base splitpath function which calls splitpath on the normpath of self
+        """
+        if self.resource:
+            path_to_split = self.path_class(self.drive + self.project + ':/' + self.resource)
+        else:
+            path_to_split = self.path_class(self.drive + self.project + ':/')
+
+        parent, child = self.path_module.split(path_to_split)
+        return self.path_class(parent), child